feat: enhance MATLAB parity with precision improvements and edge budget adjustments
- Implemented "Measure 3" tightening with a hard distance cutoff for watershed expansion.
- Updated default edge influence sigma to $2/3$ for alignment with MATLAB.
- Replaced `np.isclose` with bit-exact equality for tie-breaking in the frontier priority queue.
- Enforced `float64` precision across critical functions, removing `float32` casts.
- Corrected edge budget handling to ensure high-degree hubs can initiate exploratory traces accurately.

+++ tests/unit/core/test_global_watershed_comprehensive.py
@@ -779,7 +779,7 @@ def test_finalize_edge_trace_samples_correctly():
     assert trace.dtype == np.float32
 
     # Verify energy sampling
-    assert energy_trace.dtype == np.float32
+    assert energy_trace.dtype == np.float64
     assert len(energy_trace) == 3
 
     # Verify scale sampling
